Lots of fixes. Plugin mounting on Quest

+++ bindings/unreal/ShowtimeUnreal/Source/ShowtimeRuntime/Private/ShowtimeClient.cpp
@@ -12,12 +12,12 @@ UShowtimeClient::UShowtimeClient()
 
 void UShowtimeClient::Init()
 {
-	client->init(TCHAR_TO_UTF8(*ClientName), true);
+	if (client) client->init(TCHAR_TO_UTF8(*ClientName), true);
 }
 
 void UShowtimeClient::JoinServerByName(const FString& name)
 {
-	client->auto_join_by_name_async(TCHAR_TO_UTF8(*name));
+	if (client) client->auto_join_by_name_async(TCHAR_TO_UTF8(*name));
 }
 
 //void UShowtimeClient::JoinServerByAddress(const FIPv4Address& address)
@@ -28,7 +28,12 @@ void UShowtimeClient::JoinServerByName(const FString& name)
 
 void UShowtimeClient::LeaveServer()
 {
-	client->leave();
+	if (client) client->leave();
+}
+
+bool UShowtimeClient::IsConnected()
+{
+	return (client) ? client->is_connected() : false;
 }
 
 void UShowtimeClient::BeginPlay()
@@ -40,14 +45,15 @@ void UShowtimeClient::BeginPlay()
 
 void UShowtimeClient::BeginDestroy()
 {
-	client->destroy();
+	Super::BeginDestroy();
+	if(client) client->destroy();
 }
 
 void UShowtimeClient::EndPlay(const EEndPlayReason::Type EndPlayReason)
 {
 	Super::EndPlay(EndPlayReason);
 	RemoveEvents();
-	client->destroy();
+	if (client) client->destroy();
 }
 
 TSharedPtr<ShowtimeClient>& UShowtimeClient::Handle()
@@ -56,12 +62,18 @@ TSharedPtr<ShowtimeClient>& UShowtimeClient::Handle()
 }
 
 void UShowtimeClient::AttachEvents(){
+	if (!client)
+		return;
+
 	client_adaptor = std::make_shared<ClientAdaptors>(this);
 	client->add_connection_adaptor(client_adaptor);
 	client->add_log_adaptor(client_adaptor);
 }
 
 void UShowtimeClient::RemoveEvents(){
+	if (!client)
+		return;
+
 	client->remove_connection_adaptor(client_adaptor);
 	client->remove_log_adaptor(client_adaptor);
 }
@@ -69,5 +81,5 @@ void UShowtimeClient::RemoveEvents(){
 void UShowtimeClient::TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction)
 {
 	Super::TickComponent(DeltaTime, TickType, ThisTickFunction);
-	client->poll_once();
+	if(client) client->poll_once();
 }
